rename the modules metric, model_util, and script
