--- dattri/metrics/__init__.py
@@ -1,1 +1,0 @@
-"""dattri.metrics for some metric functions on data attribution."""
