+++ dattri/model_util/__init__.py
@@ -1,0 +1,1 @@
+"""Helper functions that are used to manipulate Pytorch models."""
