--- dattri/model_utils/__init__.py
@@ -1,1 +1,0 @@
-"""dattri.model_utils for some helper function on nn.module."""
