+++ docs/source/api/benchmark.rst
@@ -1,14 +1,14 @@
 Benchmark Functions
 =======
-.. autofunction:: dattri.model_utils.retrain.retrain_loo
-.. autofunction:: dattri.model_utils.retrain.retrain_lds
+.. autofunction:: dattri.model_util.retrain.retrain_loo
+.. autofunction:: dattri.model_util.retrain.retrain_lds
 
-.. autofunction:: dattri.metrics.ground_truth.calculate_loo_ground_truth
-.. autofunction:: dattri.metrics.ground_truth.calculate_lds_ground_truth
+.. autofunction:: dattri.metric.calculate_loo_ground_truth
+.. autofunction:: dattri.metric.calculate_lds_ground_truth
 .. autofunction:: dattri.benchmark.utils.flip_label
 
-.. autofunction:: dattri.metrics.metrics.lds
-.. autofunction:: dattri.metrics.metrics.loo_corr
-.. autofunction:: dattri.metrics.metrics.mislabel_detection_auc
+.. autofunction:: dattri.metric.lds
+.. autofunction:: dattri.metric.loo_corr
+.. autofunction:: dattri.metric.mislabel_detection_auc
 
 .. autofunction:: dattri.benchmark.load.load_benchmark
